backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.30.1
pydantic==2.7.1
python-dotenv==1.0.1
bcrypt==4.1.2
pyjwt==2.8.0
email-validator==2.1.1
